Fixed a fatal bug while moving a file using renaming
Using fs.rename can only be done on the same partition. While using this package with docker, the file system is not considered as the same one in some cases which leads to an exception and crashed the meteor runtime. I fixed it thanks to: http://stackoverflow.com/questions/4568689/how-do-i-move-file-a-to-a-different-partition-or-device-in-node-js